added a bunch of acl tests in the forum

+++ inyoka/forum/views.py
@@ -38,7 +38,7 @@ from inyoka.forum.forms import NewPostForm, NewTopicForm, SplitTopicForm, \
                                AddAttachmentForm, EditPostForm, AddPollForm, \
                                MoveTopicForm, ReportTopicForm, ReportListForm
 from inyoka.forum.acl import filter_invisible, get_forum_privileges, \
-                             have_privilege
+                             have_privilege, get_privileges
 
 
 _legacy_forum_re = re.compile(r'^/forum/(\d+)(?:/(\d+))?/?$')
@@ -200,8 +200,6 @@ def viewtopic(request, topic_slug, page=1):
     }
 
 
-@check_login(message=u'Du musst eingeloggt sein, um einen Beitrag zu '
-                     u'schreiben')
 @templated('forum/edit.html')
 def newpost(request, topic_slug=None, quote_id=None):
     """
@@ -258,6 +256,8 @@ def newpost(request, topic_slug=None, quote_id=None):
                           % att_name)
 
         elif 'delete_attachment' in request.POST:
+            if 'upload' not in privileges or 'delete' not in privileges:
+                return abort_access_denied(request)
             id = int(request.POST['delete_attachment'])
             att_ids.remove(id)
             att = filter(lambda a: a.id == id, attachments)[0]
@@ -308,8 +308,6 @@ def newpost(request, topic_slug=None, quote_id=None):
     }
 
 
-@check_login(message=u'Du musst eingeloggt sein, um ein neues Thema zu '
-                     u'erstellen')
 @templated('forum/edit.html')
 def newtopic(request, slug=None, article=None):
     """
@@ -392,6 +390,8 @@ def newtopic(request, slug=None, article=None):
                           % att_name)
 
         elif 'delete_attachment' in request.POST:
+            if 'upload' not in privileges or 'delete' not in privileges:
+                return abort_access_denied(request)
             id = int(request.POST['delete_attachment'])
             att_ids.remove(id)
             att = filter(lambda a: a.id == id, attachments)[0]
@@ -401,6 +401,8 @@ def newtopic(request, slug=None, article=None):
 
         elif 'add_poll' in request.POST:
             # the user added a new poll
+            if 'create_poll' not in privileges:
+                return abort_access_denied(request)
             poll_form = AddPollForm(request.POST)
             if poll_form.is_valid():
                 d = poll_form.cleaned_data
@@ -415,10 +417,14 @@ def newtopic(request, slug=None, article=None):
                 options = ['', '']
 
         elif 'add_option' in request.POST:
+            if 'create_poll' not in privileges:
+                return abort_access_denied(request)
             poll_form = AddPollForm(request.POST)
             options.append('')
 
         elif 'delete_poll' in request.POST:
+            if 'create_poll' not in privileges or not 'delete' in privileges:
+                return abort_access_denied(request)
             # the user deleted a poll
             poll_id = int(request.POST['delete_poll'])
             poll = filter(lambda v: v.id == poll_id, polls)[0]
@@ -496,7 +502,6 @@ def newtopic(request, slug=None, article=None):
     }
 
 
-@check_login(message=u'Du musst eingeloggt sein, um deine Beiträge zu ändern')
 @templated('forum/edit.html')
 def edit(request, post_id):
     """
@@ -505,6 +510,9 @@ def edit(request, post_id):
     polls and the options (e.g. sticky) of the topic.
     """
     post = Post.objects.get(id=post_id)
+    privileges = get_forum_privileges(request.user, post.topic.forum)
+    if 'edit' not in privileges:
+        return abort_access_denied(request)
     is_first_post = post.topic.first_post_id == post.id
     attach_form = AddAttachmentForm()
     if is_first_post:
@@ -515,6 +523,8 @@ def edit(request, post_id):
     if request.method == 'POST':
         form = EditPostForm(request.POST)
         if 'attach' in request.POST:
+            if 'upload' not in privileges:
+                return abort_access_denied(request)
             attach_form = AddAttachmentForm(request.POST, request.FILES)
             if attach_form.is_valid():
                 d = attach_form.cleaned_data
@@ -539,6 +549,8 @@ def edit(request, post_id):
                           % att_name)
 
         elif 'delete_attachment' in request.POST:
+            if 'upload' not in privileges or 'delete' not in privileges:
+                return abort_access_denied(request)
             id = int(request.POST['delete_attachment'])
             att = filter(lambda a: a.id == id, attachments)[0]
             att.delete()
@@ -546,6 +558,8 @@ def edit(request, post_id):
             flash(u'Der Anhang „%s“ wurde gelöscht.' % att.name)
 
         elif is_first_post and 'add_poll' in request.POST:
+            if 'create_poll' not in privileges:
+                return abort_access_denied(request)
             # the user added a new poll
             poll_form = AddPollForm(request.POST)
             if poll_form.is_valid():
@@ -564,10 +578,14 @@ def edit(request, post_id):
                 options = ['', '']
 
         elif is_first_post and 'add_option' in request.POST:
+            if 'create_poll' not in privileges:
+                return abort_access_denied(request)
             poll_form = AddPollForm(request.POST)
             options.append('')
 
         elif is_first_post and 'delete_poll' in request.POST:
+            if 'create_poll' not in privileges or not 'delete' in privileges:
+                return abort_access_denied(request)
             # the user deleted a poll
             poll_id = int(request.POST['delete_poll'])
             poll = filter(lambda p: p.id == poll_id, polls)[0]
@@ -613,6 +631,8 @@ def edit(request, post_id):
 def change_status(request, topic_slug, solved=None, locked=None):
     """Change the status of a topic and redirect to it"""
     t = Topic.objects.get(slug=topic_slug)
+    if not have_privilege(request.user, f.forum, 'read'):
+        abort_access_denied(request)
     if solved is not None:
         t.solved = solved
         flash(u'Das Thema wurde als %s markiert' % (solved and u'gelöst' or \
@@ -632,6 +652,8 @@ def subscribe_topic(request, topic_slug):
     If there isn't such a subscription, a new one is created.
     """
     t = Topic.objects.get(slug=topic_slug)
+    if not have_privilege(request.user, t.forum, 'read'):
+        return abort_access_denied(request)
     try:
         s = Subscription.objects.get(user=request.user, topic=t)
     except Subscription.DoesNotExist:
@@ -652,6 +674,8 @@ def subscribe_topic(request, topic_slug):
 def report(request, topic_slug):
     """Change the report_status of a topic and redirect to it"""
     t = Topic.objects.get(slug=topic_slug)
+    if not have_privilege(request.user, t.forum, 'read'):
+        return abort_access_denied(request)
     if t.reported:
         flash(u'Dieses Thema wurde bereits gemeldet; die Moderatoren werden '
               u'sich in Kürze darum kümmern.')
@@ -676,7 +700,6 @@ def report(request, topic_slug):
     }
 
 
-#TODO Check if user is a moderator
 @templated('forum/reportlist.html')
 def reportlist(request):
     """Get a list of all reported topics"""
@@ -693,12 +716,19 @@ def reportlist(request):
             Topic.objects.mark_as_done(d['selected'])
             topics = filter(lambda t: str(t.id) not in d['selected'], topics)
             flash(u'Die gewählten Themen wurden als bearbeitet markiert.',
-                                                                         True)
+                  True)
     else:
         form = ReportListForm()
         _add_field_choices()
+
+    privileges = get_privileges(user, forums)
+    visible_topics = []
+    for topic in topics:
+        if privileges.get(topic.forum_id, {'moderator': False})['moderator']:
+            visible_topics.append(topic)
+
     return {
-        'topics':   list(topics),
+        'topics':   visible_topics,
         'form':     form
     }
 
@@ -721,15 +751,26 @@ def movetopic(request, topic_slug):
     """Move a topic into another forum"""
     def _add_field_choices():
         """Add dynamic field choices to the move topic formular"""
-        form.fields['forum_id'].choices = [(f.id, f.name) for f in
-                            Forum.objects.get_forums().exclude(id=t.forum.id)]
+        form.fields['forum_id'].choices = [(f.id, f.name) for f in forums]
+
     t = Topic.objects.get(slug=topic_slug)
+    if not have_privilege(request.user, t.forum, 'moderate'):
+        return abort_access_denied()
+
+    forums = filter_visible(request.user, Forum.objects.get_forms()
+                            .exclude(id=t.forum.id), 'read')
+    mapping = dict((x.id, x) for x in forums)
+    if not mapping:
+        return abort_access_denied(request)
+
     if request.method == 'POST':
         form = MoveTopicForm(request.POST)
         _add_field_choices()
         if form.is_valid():
             data = form.cleaned_data
-            f = Forum.objects.get(id=data['forum_id'])
+            f = mapping.get(data['forum_id'])
+            if f is None:
+                return abort_access_denied(request)
             Topic.objects.move(t, f)
             # send a notification to the topic author to inform him about
             # the new forum.
@@ -764,6 +805,8 @@ def splittopic(request, topic_slug):
 
     t = Topic.objects.get(slug=topic_slug)
     posts = t.post_set.all()
+    if not have_privilege(request.user, t.forum, 'moderate'):
+        return abort_access_denied(request)
 
     if request.method == 'POST':
         form = SplitTopicForm(request.POST)
@@ -799,6 +842,8 @@ def hide_post(request, post_id):
     users can't see it anymore (moderators still can).
     """
     post = Post.objects.get(id=post_id)
+    if not have_privilege(request.user, post.topic.forum, 'moderate'):
+        return abort_access_denied(request)
     if post.id == post.topic.first_post.id:
         flash(u'Der erste Beitrag eines Themas darf nicht unsichtbar gemacht '
               u'werden.')
@@ -816,6 +861,8 @@ def restore_post(request, post_id):
     normal users again.
     """
     post = Post.objects.get(id=post_id)
+    if not have_privilege(request.user, post.topic.forum, 'moderate'):
+        return abort_access_denied(request)
     post.hidden = False
     post.save()
     flash(u'Der Beitrag %s wurde wieder sichtbar gemacht.' % post_id,
@@ -830,6 +877,8 @@ def delete_post(request, post_id):
     """
     # XXX: Only administrators are allowed to do this, not moderators
     post = Post.objects.get(id=post_id)
+    if not have_privilege(request.user, post.topic.forum, 'delete'):
+        return abort_access_denied(request)
     if post.id == post.topic.first_post.id:
         flash(u'Der erste Beitrag eines Themas darf nicht unsichtbar gemacht '
               u'werden.', success=False)
@@ -851,6 +900,8 @@ def hide_topic(request, topic_slug):
     users can't see it anymore (moderators still can).
     """
     topic = Topic.objects.get(slug=topic_slug)
+    if not have_privilege(request.user, topic.forum, 'moderate'):
+        return abort_access_denied(request)
     topic.hidden = True
     topic.save()
     flash(u'Das Thema „%s“ wurde unsichtbar gemacht.' % topic.title,
@@ -864,6 +915,8 @@ def restore_topic(request, topic_slug):
     normal users again.
     """
     topic = Topic.objects.get(slug=topic_slug)
+    if not have_privilege(request.user, topic.forum, 'moderate'):
+        return abort_access_denied(request)
     topic.hidden = False
     topic.save()
     flash(u'Das Thema „%s“ wurde wieder sichtbar gemacht.' % topic.title,
@@ -878,6 +931,8 @@ def delete_topic(request, topic_slug):
     """
     # XXX: Only administrators are allowed to do this, not moderators
     topic = Topic.objects.get(slug=topic_slug)
+    if not have_privilege(request.user, topic.forum, 'moderate'):
+        return abort_access_denied(request)
     if 'message-yes' in request.POST:
         topic.delete()
         flash(u'Das Thema „%s“ wurde erfolgreich gelöscht' % topic.title,
@@ -913,6 +968,8 @@ def feed(request, component='forum', slug=None, mode='short', count=25):
 
     if component == 'topic':
         topic = get_object_or_404(Topic, slug=slug)
+        if not have_privilege(request.user, topic.forum, 'read'):
+            return abort_access_denied()
         posts = topic.post_set.order_by('-pub_date')[:count]
         feed = FeedBuilder(
             title=u'ubuntuusers Thema – „%s“' % topic.title,
@@ -959,6 +1016,8 @@ def feed(request, component='forum', slug=None, mode='short', count=25):
                 rights=href('portal', 'lizenz'),
             )
 
+        if not have_privilege(request.user, forum, 'read'):
+            return abort_access_denied()
         topics = topics.order_by('-id')[:count]
 
         for topic in topics:
@@ -993,7 +1052,7 @@ def markread(request, slug=None):
     Mark either all or only the given forum as read.
     """
     user = request.user
-    if user.is_authenticated():
+    if user.is_authenticated:
         user.forum_last_read = Post.objects.get_max_id()
         user.save()
     return HttpResponseRedirect(href('forum'))
